Change SetConfiguration parameter to short allowing for -1.
The libusb set_configuration documentation says that -1 should
be used to choose non-configured state for a device. The current
SetConfiguration uses a 'byte' type which is unsigned. Changing
to 'short' to allow for negative values.  Internally when the
USB message SET_CONFIGURATION is composed the value is typed as
'short' (at least in the .NET code).

+++ LibWinUsb/IUsbDevice.cs
@@ -41,15 +41,17 @@ namespace LibUsbDotNet
     /// </example>
     public interface IUsbDevice : IUsbInterface
     {
-        /// <summary>
-        /// Sets the USB devices active configuration value. 
-        /// </summary>
-        /// <param name="config">The active configuration value. A zero value means the device is not configured and a non-zero value indicates the device is configured.</param>
-        /// <returns>True on success.</returns>
-        /// <remarks>
-        /// A USB device can have several different configurations, but only one active configuration.
-        /// </remarks>
-        bool SetConfiguration(byte config);
+		/// <summary>
+		/// Sets the USB devices active configuration value. 
+		/// </summary>
+		/// <param name="config">The active configuration value. 
+		/// A zero value means the device is not configured and a non-zero value indicates the device is configured.
+		/// According to the libusb documentation, a value of -1 should be used to indicate unconfigured device.</param>
+		/// <returns>True on success.</returns>
+		/// <remarks>
+		/// A USB device can have several different configurations, but only one active configuration.
+		/// </remarks>
+		bool SetConfiguration(short config);
 
         /// <summary>
         /// Gets the USB devices active configuration value.
